The starting value of Azimuth and Elevation are now pulled from the webservice.
When getSensorData returns, it will check if the Azimith_Motor or the Elevation_Motor sensors have a value. If they do azimuth and elevation will be set to these values if their parameters cannot be found.

Confimed to work with the sliders option. However since the tcp server does not work at the moment it is possible sending coordinates directly may not update the local value properly.